# Translation.Backend/generated/models/AudioUpload.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class AudioUpload
    {
        public string UploadId { get; set; }

        public string FileName { get; set; }

        public float FileSize { get; set; }

        public string UploadDate { get; set; }


    }
}