# Translation.Backend/generated/models/Payment.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class Payment
    {
        public string PaymentId { get; set; }

        public string UserId { get; set; }

        public float Amount { get; set; }

        public string[] LanguageOptions { get; set; }

        public string Service { get; set; }


    }
}