# Translation.Backend/generated/models/PaymentFailureError.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class PaymentFailureError : ErrorResponse
    {
        public new string ErrorCode { get; } = "PAYMENT_FAILURE";

        public new string ErrorMessage { get; } = "Payment processing failed.";


    }
}