# Translation.Backend/generated/models/RateLimiting.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class RateLimiting
    {
        public int Limit { get; set; }

        public string Window { get; set; }


    }
}