# Translation.Backend/generated/models/TranslationJob.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class TranslationJob
    {
        public string JobId { get; set; }

        public string UploadId { get; set; }

        public string Status { get; set; }

        public float Progress { get; set; }

        public string CreatedAt { get; set; }


    }
}