# servers/aspnet/generated/models/InvalidFileError.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class InvalidFileError : ErrorResponse
    {
        public new string ErrorCode { get; } = "INVALID_FILE";

        public new string ErrorMessage { get; } = "The uploaded file is invalid.";


    }
}