# servers/aspnet/generated/models/NotificationPreferences.cs
// Copyright (c) Microsoft Corporation. All rights reserved.
// Licensed under the MIT License.
// <auto-generated />

using System;
using System.Text.Json;
using System.Text.Json.Serialization;

namespace AudioTranslationService.Models.Service.Models
{

    public partial class NotificationPreferences
    {
        public string UserId { get; set; }

        public bool EmailNotifications { get; set; }


    }
}